robot will update the runfile every 2 minutes.
if the file is more than two minutes old -- and you can confirm the
timestamp by looking at the contents -- we have reason to suspect the
robot is dead. the google drive link may be down, or the indesign script
may not be running, or the computer may be off.